pyproject.toml
[project]
name = "books-uniq"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pingfury", email = "pingfury@outlook.com" }
]
dependencies = [
    "fastapi[standard]>=0.116.1",
    "chromadb>=0.4.24",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.24.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/books_uniq"]
